🧬 Migrate the test suite to the correct test module.
