+++ test/unit_test/server/application/init.py
@@ -1,9 +1,7 @@
-import json
-import os
 from abc import ABCMeta, abstractmethod
 from collections import namedtuple
 from typing import Any, Type, Union, cast
-from unittest.mock import Mock, mock_open, patch
+from unittest.mock import Mock, patch
 
 import fastapi
 import pytest
@@ -13,9 +11,7 @@ from flask import Flask
 from flask import Request as FlaskRequest
 from flask import Response as FlaskResponse
 
-from pymock_api.exceptions import FileFormatNotSupport
 from pymock_api.server.application import BaseAppServer, FastAPIServer, FlaskServer
-from pymock_api.server.application.response import HTTPResponse as _HTTPResponse
 
 MockerModule = namedtuple("MockerModule", ["module_path", "return_value"])
 
@@ -157,51 +153,3 @@ class TestFastAPIServer(AppServerTestSpec):
     @property
     def _expected_response_type(self) -> Type[FastAPIResponse]:
         return FastAPIResponse
-
-
-class TestInnerHTTPResponse:
-    @pytest.fixture(scope="function")
-    def http_resp(self) -> Type[_HTTPResponse]:
-        return _HTTPResponse
-
-    def test_response_with_string_data(self, http_resp: Type[_HTTPResponse]):
-        resp_data = http_resp.generate(data=_General_String_Value)
-        assert resp_data == _General_String_Value, ""
-
-    def test_response_with_json_format_string_data(self, http_resp: Type[_HTTPResponse]):
-        json_content_str = json.dumps(_Json_File_Content)
-        resp_data = http_resp.generate(data=json_content_str)
-        assert resp_data == json.loads(json_content_str), ""
-
-    def test_response_with_json_file_name(self, http_resp: Type[_HTTPResponse]):
-        with patch.object(os.path, "exists", return_value=True) as os_path_exists:
-            json_content_str = json.dumps(_Json_File_Content)
-            with patch("builtins.open", mock_open(read_data=json_content_str)) as mock_file_stream:
-                # Run target function to test
-                resp_data = http_resp.generate(data=_Json_File_Name)
-
-                # Verify result
-                os_path_exists.assert_called_once_with(_Json_File_Name)
-                mock_file_stream.assert_called_once_with(_Json_File_Name, "r", encoding="utf-8")
-                assert open(_Json_File_Name).read() == json_content_str, ""
-                assert resp_data == _Json_File_Content, ""
-
-    def test_response_with_not_exist_json_file_name(self, http_resp: Type[_HTTPResponse]):
-        with patch("builtins.open", mock_open(read_data=None)) as mock_file_stream:
-            with pytest.raises(FileNotFoundError) as exc_info:
-                # Run target function to test
-                http_resp.generate(data=_Not_Exist_File_Name)
-                # Verify result
-                expected_err_msg = f"The target configuration file {_Not_Exist_File_Name} doesn't exist."
-                assert str(exc_info) == expected_err_msg, f"The error message should be same as '{expected_err_msg}'."
-                mock_file_stream.assert_not_called()
-
-    def test_response_with_not_json_file_name(self, http_resp: Type[_HTTPResponse]):
-        with patch("builtins.open", mock_open(read_data=None)) as mock_file_stream:
-            with pytest.raises(FileFormatNotSupport) as exc_info:
-                # Run target function to test
-                http_resp.generate(data=_Not_Json_File_Name)
-                # Verify result
-                expected_err_msg = f"It doesn't support reading '{', '.join(http_resp.valid_file_format)}' format file."
-                assert str(exc_info) == expected_err_msg, f"The error message should be same as '{expected_err_msg}'."
-                mock_file_stream.assert_not_called()
